requirements.txt
streamlit==1.30.0
openai==1.4.0
python-dotenv==1.0.0
tqdm==4.66.1
PyPDF2==3.0.1
nltk==3.8.1
